Revert "Lucas/sensible s3 load model"
