+++ pyproject.toml
@@ -30,5 +30,4 @@ dependencies = [
     "wandb",
     "python-dotenv",
     "tqdm",
-    "boto3"
 ]
